Split up persistence package
