+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/datatypes/ChunkPositionPersistentDataType.kt
@@ -1,11 +1,4 @@
-package io.github.pylonmc.pylon.core.persistence.datatypes
-
-import io.github.pylonmc.pylon.core.pluginInstance
-import io.github.pylonmc.pylon.core.util.position.ChunkPosition
-import org.bukkit.NamespacedKey
-import org.bukkit.persistence.PersistentDataAdapterContext
-import org.bukkit.persistence.PersistentDataContainer
-import org.bukkit.persistence.PersistentDataType
+package io.github.pylonmc.pylon.core.datatypes
 
 object ChunkPositionPersistentDataType : PersistentDataType<PersistentDataContainer, ChunkPosition> {
     val xKey = NamespacedKey(pluginInstance, "x")
