+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/datatypes/MapPersistentDataType.kt
@@ -1,10 +1,4 @@
-package io.github.pylonmc.pylon.core.persistence.datatypes
-
-import io.github.pylonmc.pylon.core.pluginInstance
-import org.bukkit.NamespacedKey
-import org.bukkit.persistence.PersistentDataAdapterContext
-import org.bukkit.persistence.PersistentDataContainer
-import org.bukkit.persistence.PersistentDataType
+package io.github.pylonmc.pylon.core.datatypes
 
 class MapPersistentDataType<KP : Any, KC : Any, VP : Any, VC : Any>(
     private val keyType: PersistentDataType<KP, KC>,
