+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/datatypes/WorldPersistentDataType.kt
@@ -1,9 +1,4 @@
-package io.github.pylonmc.pylon.core.persistence.datatypes
-
-import org.bukkit.Bukkit
-import org.bukkit.World
-import org.bukkit.persistence.PersistentDataAdapterContext
-import org.bukkit.persistence.PersistentDataType
+package io.github.pylonmc.pylon.core.datatypes
 
 object WorldPersistentDataType : PersistentDataType<LongArray, World> {
     override fun getPrimitiveType(): Class<LongArray> = LongArray::class.java
