+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/debug/DebugWaxedWeatheredCutCopperStairs.kt
@@ -7,7 +7,7 @@ import io.github.pylonmc.pylon.core.item.LoreBuilder
 import io.github.pylonmc.pylon.core.item.PylonItem
 import io.github.pylonmc.pylon.core.item.PylonItemSchema
 import io.github.pylonmc.pylon.core.item.base.BlockInteractor
-import io.github.pylonmc.pylon.core.persistence.blockstorage.BlockStorage
+import io.github.pylonmc.pylon.core.block.BlockStorage
 import io.github.pylonmc.pylon.core.util.plus
 import io.github.pylonmc.pylon.core.util.pylonKey
 import io.papermc.paper.datacomponent.DataComponentTypes
