+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/test/block/BlockStorageMissingSchemaTest.java
@@ -6,8 +6,8 @@ import io.github.pylonmc.pylon.core.block.PylonBlockSchema;
 import io.github.pylonmc.pylon.core.event.PylonBlockLoadEvent;
 import io.github.pylonmc.pylon.core.event.PylonChunkBlocksLoadEvent;
 import io.github.pylonmc.pylon.core.event.PylonChunkBlocksUnloadEvent;
-import io.github.pylonmc.pylon.core.persistence.blockstorage.BlockStorage;
-import io.github.pylonmc.pylon.core.persistence.blockstorage.PhantomBlock;
+import io.github.pylonmc.pylon.core.block.BlockStorage;
+import io.github.pylonmc.pylon.core.block.PhantomBlock;
 import io.github.pylonmc.pylon.core.registry.PylonRegistry;
 import io.github.pylonmc.pylon.test.PylonTest;
 import io.github.pylonmc.pylon.test.TestUtil;
